refactor: extract shared MCP test server logic into base class
